backend/requirements.txt
google-genai==1.11.0
pydantic==2.11.3
boto3==1.37.36
